+++ tests/drum/unit/test_structured_input_read_utils.py
@@ -4,11 +4,13 @@ All rights reserved.
 This is proprietary source code of DataRobot, Inc. and its affiliates.
 Released under the terms of DataRobot Tool and Utility Agreement.
 """
+import os
 import tempfile
 
-
+import pandas as pd
 import numpy as np
-
+import pyarrow
+from pandas._testing import assert_frame_equal
 
 from datarobot_drum.drum.utils import StructuredInputReadUtils
 
@@ -30,3 +32,37 @@ class TestStructuredInputReadUtils(object):
 
         assert np.isnan(X["data"][5])
         assert X.shape[0] == 7
+
+    def test_read_structured_input_arrow_csv_na_consistency(self, tmp_path):
+        """
+        Test that N/A values (None, numpy.nan) are handled consistently when using
+        CSV vs Arrow as a prediction payload format.
+        1. Make CSV and Arrow prediction payloads from the same dataframe
+        2. Read both payloads
+        3. Assert the resulting dataframes are equal
+        """
+
+        # arrange
+        df = pd.DataFrame({"col_int": [1, np.nan, None], "col_obj": ["a", np.nan, None]})
+
+        csv_filename = os.path.join(tmp_path, "X.csv")
+        with open(csv_filename, "w") as f:
+            f.write(df.to_csv(index=False))
+
+        arrow_filename = os.path.join(tmp_path, "X.arrow")
+        with open(arrow_filename, "wb") as f:
+            f.write(pyarrow.ipc.serialize_pandas(df).to_pybytes())
+
+        # act
+        csv_df = StructuredInputReadUtils.read_structured_input_file_as_df(csv_filename)
+        arrow_df = StructuredInputReadUtils.read_structured_input_file_as_df(arrow_filename)
+
+        # assert
+        is_nan = lambda x: isinstance(x, float) and np.isnan(x)
+        is_none = lambda x: x is None
+
+        assert_frame_equal(csv_df, arrow_df)
+        # `assert_frame_equal` doesn't make a difference between None and np.nan.
+        # To do an exact comparison, compare None and np.nan "masks".
+        assert_frame_equal(csv_df.applymap(is_nan), arrow_df.applymap(is_nan))
+        assert_frame_equal(csv_df.applymap(is_none), arrow_df.applymap(is_none))
